feat: notification ui without buttons
