+++ shell/crates/notifications_drawer/src/components/mod.rs
@@ -3,7 +3,19 @@ use bevy::{
     ecs::{relationship::RelatedSpawner, spawn::SpawnWith},
     prelude::*,
 };
+
 pub mod cards;
+pub mod app_name_row;
+pub mod buttons;
+pub mod icon;
+pub mod surface;
+
+pub use cards::*;
+pub use app_name_row::*;
+pub use buttons::*;
+pub use icon::*;
+pub use surface::*;
+
 #[derive(Component)]
 pub struct NotificationButton {
     pub id: u32,
