+++ shell/crates/notifications_drawer/src/lib.rs
@@ -2,18 +2,11 @@ use bevy::prelude::*;
 
 pub mod components;
 pub mod systems;
+pub mod plugin;
 mod ui;
 
-use systems::*;
+pub use plugin::NotificationDrawerPlugin;
 
-// pub struct NotificationsDrawerPlugin;
-// impl Plugin for NotificationsDrawerPlugin {
-//     fn build(&self, app: &mut App) {
-//         app.add_systems(Startup, setup);
-//         app.add_systems(Update, (button_system, exit_on_esc));
-//     }
-// }
-
-// pub mod prelude {
-//     pub use crate::NotificationsDrawerPlugin;
-// }
+pub mod prelude {
+    pub use crate::NotificationDrawerPlugin;
+}
