+++ shell/crates/notifications_drawer/src/systems/mod.rs
@@ -1,5 +1,9 @@
-mod button_system;
+pub mod button_system;
 pub mod setup;
+pub mod expiry;
+pub mod notification_processing;
 
-pub use button_system::button_system;
+pub use button_system::*;
 pub use setup::{exit_on_esc, setup};
+pub use expiry::*;
+pub use notification_processing::*;
